backend/requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
itsdangerous==2.0.1
flask-cors==3.0.10
cohere==4.37
reportlab==4.0.4
python-dotenv==0.19.2
aiohttp==3.8.5